Map.tsx にMap関数コンポーネントを移動
